+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * Add method to filter entries from a dictionary.
   * Add alternative URL Query parameter encoding implementation.
   * Add method to determine if one dictionary is a subset of another.
+  * Add method to construct dictionary by combining pairs.
 
 2.4.0, 26-June-2014